+++ src/frontend/src/app/components/subjects/subject-document/sub-doc-edit/sub-learn-outc/sub-learn-outc.component.ts
@@ -2,6 +2,7 @@ import { Component, Input, OnInit } from '@angular/core';
 import { LearningOutcome } from 'src/app/core/models/learning-outcome/learning-outcome';
 import { LearningOutcomeEvaluation } from 'src/app/core/models/subject/learning-outcome-evaluation';
 import { Subject } from 'src/app/core/models/subject/subject';
+import { AlertService } from 'src/app/services/alerts/alert.service';
 import { LearningOutcomeService } from 'src/app/services/learning-outcome/learning-outcome.service';
 
 @Component({
@@ -21,14 +22,15 @@ export class SubLearnOutcComponent implements OnInit {
   
   learningOutcomes: LearningOutcome[] = [];
   
-  constructor(private readonly learningOutcomeService: LearningOutcomeService) { }
+  constructor(private readonly learningOutcomeService: LearningOutcomeService,
+    private alerts: AlertService) { }
 
   ngOnInit(): void {
   }
 
   loadLearningOutcomes() {
     this.learningOutcomes = [];
-    this.learningOutcomeService.getLatest(this._document.fieldOfStudy.code, this._document.academicYear).subscribe(outcomes => {
+    this.learningOutcomeService.getLatestReadOnly(this._document.fieldOfStudy.code, this._document.academicYear).subscribe(outcomes => {
       this.learningOutcomes = outcomes?.learningOutcomes ?? [];
     })
   }
@@ -47,8 +49,19 @@ export class SubLearnOutcComponent implements OnInit {
   }
 
   save(learn: LearningOutcomeEvaluation) {
-    this.delete();
+    if (this.checkLoIsUnique(learn)) {
+      this.delete();
     this._document.learningOutcomeEvaluations.push(learn);
+    }
+    else {
+      this.alerts.showCustomErrorMessage('Literatura o podanym numerze ISBN już istnieje!');
+    }
   }
 
+  checkLoIsUnique(lo: LearningOutcomeEvaluation): boolean {
+    if(this._document.learningOutcomeEvaluations.find(l => l.learningOutcomeSymbol === lo.learningOutcomeSymbol) && this.selected?.learningOutcomeSymbol !== lo.learningOutcomeSymbol) {
+      return false;
+    }
+    return true;
+  }
 }
