+++ src/frontend/src/app/components/subjects/subject-document/sub-doc-edit/sub-literature/sub-literature.component.ts
@@ -1,6 +1,7 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { Literature } from 'src/app/core/models/subject/literature';
 import { Subject } from 'src/app/core/models/subject/subject';
+import { AlertService } from 'src/app/services/alerts/alert.service';
 
 @Component({
   selector: 'app-sub-literature',
@@ -12,7 +13,7 @@ export class SubLiteratureComponent implements OnInit {
   @Input() document: Subject = new Subject();
   selected: Literature | null = null;
   
-  constructor() { }
+  constructor(private alerts: AlertService) { }
 
   ngOnInit(): void {
   }
@@ -31,7 +32,19 @@ export class SubLiteratureComponent implements OnInit {
   }
 
   save(lit: Literature) {
-    this.delete();
-    this.document.literature.push(lit);
+    if (this.checkIsbnIsUnique(lit)) {
+      this.delete();
+      this.document.literature.push(lit);
+    }
+    else {
+      this.alerts.showCustomErrorMessage('Literatura o podanym numerze ISBN już istnieje!');
+    }
+  }
+
+  checkIsbnIsUnique(lit: Literature): boolean {
+    if(this.document.literature.find(l => l.isbn === lit.isbn) && this.selected?.isbn !== lit.isbn) {
+      return false;
+    }
+    return true;
   }
 }
